requirements.txt
PyJWT==2.0.1
requests==2.25.1